+++ CHANGELOG.md
@@ -21,8 +21,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+<<<<<<< HEAD
 - New component nc-button #17
+- New component nc-slideshow #6
 - New component nc-modal #4
 - New component nc-list #11
 - Added package.json configuration #8
 - New component nc-loader #6
+
+### Changed
+- Component nc-list to customize item #96